requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
azure-functions==1.17.0
blinker==1.6.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
more-itertools==10.1.0
multidict==6.0.4
numpy==1.26.0
openai==0.28.0
openapi-core==0.18.1
openapi-schema-validator==0.6.1
openapi-spec-validator==0.6.0
packaging==23.1
parse==1.19.1
pathable==0.4.3
prance==23.6.21.0
pydantic==1.10.12
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.10.3
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
semantic-kernel==0.3.13.dev0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
yarl==1.9.2
